Add required API endpoints to api.php
